Separate examples test for hybridization solver
